Feat(SCRUM-14): 예외 로그 개선
