+++ src/main/java/io/ourfit/api/global/exception/GlobalExceptionHandler.java
@@ -6,11 +6,13 @@ import static org.springframework.http.HttpStatus.METHOD_NOT_ALLOWED;
 import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 import io.ourfit.api.global.data.dto.ErrorResponse;
+import io.ourfit.api.global.web.hook.WebhookService;
 import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.Valid;
 import java.util.*;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.http.HttpHeaders;
@@ -27,14 +29,18 @@ import org.springframework.web.method.annotation.HandlerMethodValidationExceptio
 import org.springframework.web.servlet.NoHandlerFoundException;
 
 /** 전역 예외 처리 클래스 */
+@Slf4j
 @RestControllerAdvice
 @RequiredArgsConstructor
 public class GlobalExceptionHandler {
 
+  private final ExceptionLogFormatter formatter;
   private final MessageSource customMessageSource;
+  private final WebhookService webhookService;
 
   @ExceptionHandler(Exception.class)
-  protected ResponseEntity<ErrorResponse> handleAllUncaughtException(Exception ignored) {
+  protected ResponseEntity<ErrorResponse> handleAllUncaughtException(Exception ex) {
+    this.logException(ex, true);
     return ResponseEntity.internalServerError()
         .body(ErrorResponse.internalServerError(this.resolveMessage(INTERNAL_SERVER_ERROR)));
   }
@@ -44,20 +50,23 @@ public class GlobalExceptionHandler {
     HandlerMethodValidationException.class,
     HttpMessageNotReadableException.class
   })
-  protected ResponseEntity<ErrorResponse> handleHttpMessageNotReadableException(Exception ignored) {
+  protected ResponseEntity<ErrorResponse> handleHttpMessageNotReadableException(Exception ex) {
+    this.logException(ex, false);
     return ResponseEntity.badRequest()
         .body(ErrorResponse.badRequest(this.resolveMessage(INVALID_PARAMETER)));
   }
 
   @ExceptionHandler(AuthenticationException.class)
   protected ResponseEntity<ErrorResponse> handleAuthenticationException(
-      AuthenticationException ignored) {
+      AuthenticationException ex) {
+    this.logException(ex, false);
     return ResponseEntity.status(UNAUTHORIZED)
         .body(ErrorResponse.unauthorized(this.resolveMessage(ApiExceptionType.UNAUTHORIZED)));
   }
 
   @ExceptionHandler(OurfitApiException.class)
   protected ResponseEntity<ErrorResponse> handleLocatApiException(OurfitApiException ex) {
+    this.logException(ex, false);
     return ResponseEntity.status(ex.getHttpStatus())
         .body(ErrorResponse.fromException(ex, this.resolveMessage(ex.getApiExceptionType())));
   }
@@ -76,6 +85,7 @@ public class GlobalExceptionHandler {
             "io.ourfit.api.exception.UNSUPPORTED_METHOD.message",
             new Object[] {ex.getMethod(), supportedMethods},
             LocaleContextHolder.getLocale());
+    this.logException(ex, false);
     return ResponseEntity.status(METHOD_NOT_ALLOWED)
         .header(HttpHeaders.ALLOW, supportedMethods)
         .body(ErrorResponse.methodNotAllowed(message));
@@ -91,7 +101,7 @@ public class GlobalExceptionHandler {
                 fieldError ->
                     "%s - %s.".formatted(fieldError.getField(), fieldError.getDefaultMessage()))
             .collect(Collectors.joining());
-
+    this.logException(ex, false);
     return ResponseEntity.badRequest().body(ErrorResponse.badRequest(message));
   }
 
@@ -108,14 +118,25 @@ public class GlobalExceptionHandler {
 						"""
                         .formatted(violation.getPropertyPath(), violation.getMessage()))
             .collect(Collectors.joining());
+    this.logException(ex, false);
     return ResponseEntity.badRequest().body(ErrorResponse.badRequest(message));
   }
 
   @ExceptionHandler(NoHandlerFoundException.class)
   protected ResponseEntity<ErrorResponse> handleNoHandlerFoundException(
       NoHandlerFoundException ex) {
+    this.formatter.doFormat(ex);
     return ResponseEntity.status(ex.getStatusCode())
         .body(ErrorResponse.notFound(this.resolveMessage(ApiExceptionType.NOT_FOUND_API_ENDPOINT)));
+  }
+
+  private void logException(Exception ex, boolean sendWebhook) {
+    String logMessage = this.formatter.doFormat(ex);
+
+    log.error(logMessage);
+    if (sendWebhook) {
+      this.webhookService.send(logMessage);
+    }
   }
 
   private String resolveMessage(ApiExceptionType exceptionType) {
