+++ src/main/java/io/ourfit/api/global/web/hook/WebhookService.java
@@ -1,0 +1,6 @@
+package io.ourfit.api.global.web.hook;
+
+public interface WebhookService {
+
+  void send(String message);
+}
